+++ persistent/__init__.py
@@ -28,38 +28,22 @@ __all__ = [
     'PickleCache',
     'TimeStamp',
 ]
-from persistent._compat import PURE_PYTHON
-from persistent.interfaces import IPersistent
-from persistent.interfaces import IPickleCache
 
-import persistent.timestamp as TimeStamp
+# Take care not to shadow the module names
+from persistent import interfaces as _interfaces
+from persistent import timestamp as _timestamp
+from persistent import persistence as _persistence
+from persistent import picklecache as _picklecache
 
-from persistent import persistence as pyPersistence
-from persistent import picklecache as pyPickleCache
-
-try:
-    # Be careful not to shadow the modules
-    from persistent import cPersistence as _cPersistence
-    from persistent import cPickleCache as _cPickleCache
-except ImportError: # pragma: no cover
-    _cPersistence = None
-    _cPickleCache = None
-else:
-    # Make an interface declaration for Persistent
-    # Note that the Python version already does this.
-    from zope.interface import classImplements
-    classImplements(_cPersistence.Persistent, IPersistent)
-    classImplements(_cPickleCache.PickleCache, IPickleCache)
-
-
-_persistence = pyPersistence if PURE_PYTHON or _cPersistence is None else _cPersistence
-_picklecache = pyPickleCache if PURE_PYTHON or _cPickleCache is None else _cPickleCache
-
+IPersistent = _interfaces.IPersistent
 Persistent = _persistence.Persistent
-GHOST = _persistence.GHOST
-UPTODATE = _persistence.UPTODATE
-CHANGED = _persistence.CHANGED
-STICKY = _persistence.STICKY
+GHOST = _interfaces.GHOST
+UPTODATE = _interfaces.UPTODATE
+CHANGED = _interfaces.CHANGED
+STICKY = _interfaces.STICKY
 PickleCache = _picklecache.PickleCache
+
+# BWC for TimeStamp.
+TimeStamp = _timestamp
 
 sys.modules['persistent.TimeStamp'] = sys.modules['persistent.timestamp']