+++ persistent/_compat.h
@@ -31,7 +31,7 @@
 
 #define INT_FROM_LONG(x) PyLong_FromLong(x)
 #define INT_CHECK(x) PyLong_Check(x)
-#define INT_AS_LONG(x) PyLong_AS_LONG(x)
+#define INT_AS_LONG(x) PyLong_AsLong(x)
 #define CAPI_CAPSULE_NAME "persistent.cPersistence.CAPI"
 
 #else